more config. Prevent static variables from being optimized out.
Rust has the #[used] attribute to force the compiler to emit static
items in the object file. But the linker can still optimize them out, so
we use the KEEP instruction in our linker script to force it to keep the
.device_info and .wifi_info sections.

+++ common/src/lib.rs
@@ -19,7 +19,7 @@ pub enum MessageUpdateKind {
     Text(u32),
 }
 
-type DeviceID = u8;
+type DeviceID = u32;
 type UpdateID = u32;
 
 #[derive(Clone, Copy, Serialize, Deserialize)]
